[testing] Added basic functionality tests for math mode+supported parameters. Moved Math and score mode to their own subfolders in src to stop them from spilling into eachother.
